Add trades.message_discord_id and user name columns
The message_discord_id will relate a trade to specific message
that can be updated as the trade changes. The user_1_name
and user_2_name columns will be used to display user names.
Because names can change, this does not belong on the user
table but rather the trades table.