Add CLT::Repos and CLT::AddSpawn
These are untested command-line tools. However, CLT::Repos might
eventually be migrated into the rest of the codebase because
of how convenient it is.

CLT::AddSpawn is just an easy way for a developer to give a user
a Pokemon spawn of their choice. For development purpses only,
of course!