Refactor abstract base commands to use Pokecord::Repos
This simplifies concrete child classes, since they don't need to
initialize any additional repositories.

It is possible that this will slow down the speed at which user
commands are processed. To that end, I timed running the entire
rspec suite before and after this change. Before, the suite took
2.89 seconds to run. Afterwards, the suite took 2.96 seconds to
run.

As a difference of 0.07 seconds (for the entire test suite) is
pretty small to human users, I am initially declaring this
speed sacrifice worth the development quality-of-life change. If
user command slowdowns become a problem, I can revisit this
commit and do more serious benchmarking.